# backend/api.yaml
openapi: 3.0.0
info:
  title: Share2Pay API
  version: 1.0.0
servers:
  - url: https://share2paybackend.vercel.app/
tags:
  - name: User
    description: Operations related to user profile
  - name: Account
    description: Operations related to user accounts
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /:
    get:
      summary: Basic Health checkup route
      description: Basic route to check if backend is up or not
      responses:
        '200':
          description: Backend System up...
  /api/v1/user/signup:
    post:
      summary: Sign up a new user
      description: Route to sign up a new user...(regex for phoneNo is /^([+]?[\s0-9]+)?(\d{3}|[(]?[0-9]+[)])?([-]?[\s]?[0-9])+$/)
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Returns JWT token of the created user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token of the created user.
        '401':
          description: Username, phone number, or email already exists.
        '403':
          description: Unauthorized or bad authentication.
        '411':
          description: Incorrect inputs.
        '500':
          description: Internal Server Error
        '502':
          description: Error in saving data in database.
  /api/v1/user/signin:
    post:
      summary: Sign In a user
      description: Route to sign in a user...
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Returns JWT token of the created user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token of the created user.
                  username:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  phoneNumber:
                    type: string
        '400':
          description: Incorrect credentials....
        '403':
          description: Unauthorized or bad authentication.
  /api/v1/user:
    get:
      summary: Authentication check...
      description: Route to check whether user is authenticated?
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User authenticated...
        '403':
          description: Unauthorized or bad authentication.
  /api/v1/user/user:
    put:
      summary: Update user details...
      description: Route to Update user details...(regex for phoneNo is /^([+]?[\s0-9]+)?(\d{3}|[(]?[0-9]+[)])?([-]?[\s]?[0-9])+$/)
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Data Updated Succesfully...
        '403':
          description: Unauthorized or bad authentication.
        '411':
          description: Wrong inputs/Password cannot be updated from here.
  /api/v1/user/searchUsers:
    get:
      summary: Search other users..
      description: Route to see/filter all other users...
      tags: [User]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: filter
          schema:
            type: string
      responses:
        '200':
          description: Returns the list of users matching the search...
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: array
                    description: The list of users.
        '403':
          description: Unauthorized or bad authentication.
  /api/v1/account/:
    get:
      summary: Basic Health checkup for account router
      description: Basic route to check if backend is up or not
      tags: [Account]
      responses:
        '200':
          description: Backend System up... 
  /api/v1/account/balance:
    get:
      summary: Fetch balance of user
      description: Route to check the balance of the currently logged-in user.
      tags: [Account]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Returns the balance of the logged-in user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    description: The balance of the logged-in user.
        '403':
          description: Unauthorized or bad authentication.
        '500':
          description: Internal Server Error
  /api/v1/account/transfer:
    post:
      summary: Transfer balance to other user
      description: Route to transfer balance to another user...
      tags: [Account]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                toUserId:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Transfer successful
        '400':
          description: Insufficient balance or invalid request
        '411':
          description: Invalid receiver account
        '403':
          description: Unauthorized or Bad authentication
        '500':
          description: Internal Server Error